[test build] improve cpu time by 22%
this is a simple trick:
we do not immediately write every entry to the socket. instead, we append all lines to a string and then send that string, thus reducing the cpu time by a lot.
this improvement is probably even more than 22% for files with more than ~10 lines as we save *n* syscalls where n is the amount of lines